+++ src/www/fastify/Http2SecureServer.ts
@@ -21,6 +21,7 @@ export class Http2SecureServer<
 	 * @constructor
 	 * @memberof ApiServer
 	 * @since 1.0.0
+	 * @since 7.2.1 Disable fastify logs.
 	 * @author Caique Araujo <caique@piggly.com.br>
 	 */
 	constructor(
@@ -42,13 +43,8 @@ export class Http2SecureServer<
 						cert: ssl.cert,
 						key: ssl.key,
 					},
-					logger:
-						options.fastify.logger ||
-						AbstractServer.defaultLogger(
-							options.env.environment,
-							options.env.app.root_path,
-							options.env.debug,
-						),
+					/** @note There is no reason to log with fastify. */
+					logger: false,
 					trustProxy: true,
 				},
 			),